+++ sources/simulation_data.h
@@ -25,13 +25,13 @@ public:
 	
   // new data
   std::vector <Event> eventChain;
-  Event* addEvent(int time, net* net, LogicLevel state);
+  Event* addEvent(float time, net* net, LogicLevel state);
 
   // map data
-  int endTime;
-  std::map<int, Event*> newEventChain;
-  Event* addMapEvent(int time, net* net, LogicLevel state);
-  Event* addMapEvent(int time, net* net, LogicLevel state, bool delayedState);
+  float endTime;
+  std::map<float, Event*> newEventChain;
+  Event* addMapEvent(float time, net* net, LogicLevel state);
+  Event* addMapEvent(float time, net* net, LogicLevel state, bool delayedState);
 
 
   // end of map data
